+++ packages/models/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## v1.2.1
+- Add SPS30 sensor 
+- Update @sensebox/node-sketch-templater to v1.13.0
+
 ## v1.2.0
 
 - Extend `luftdatenHandler` (#578)